# services/server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leasy Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a new user",
                "description": "Create a new user with the provided information",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "full_name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "ofirgoldofir@gmail.com"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "Ofir Goldberg"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Aa123!@#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "500": {
                        "description": "An error occurred"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticate and login user based on personal number",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "ofirgoldofir@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Aa123!@#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "An error occurred while user logged in"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Logout the authenticated user by revoking the user's access token",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "500": {
                        "description": "An error occurred while logging out"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/chat/send-message/{uuid}": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Send a message to a chat",
                "description": "Sends a message to a specific chat identified by its UUID.",
                "operationId": "db372e1f498ed3dc425bbe33951ba399",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the chat",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, how are you?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/chat/messages/{uuid}": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "description": "Retrieve messages records. Supports pagination.",
                "operationId": "getChatMessages",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the chat",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with messages of the chat"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
...
    }
}